+++ mrb_doc/models/sound.rb
@@ -3,10 +3,10 @@ class Sound
   # {Sound::NotFound}.
   #
   # @example Basic usage
-  #   music = Sound.new("./assets/beep.wav")
+  #   beep = Sound.new("./assets/beep.wav")
   #
   # @example Basic usage with lower volume and higher pitch
-  #   music = Sound.new("./assets/beep.wav", volume: 0.75, pitch: 1.2)
+  #   beep = Sound.new("./assets/beep.wav", volume: 0.75, pitch: 1.2)
   #
   # @param volume [Float] A value between 0.0 and 1.0.
   # @param pitch [Float]
@@ -20,7 +20,25 @@ class Sound
     Sound.new
   end
 
+  # Unloads the {Sound} from memory.
+  #
+  # @example Basic usage
+  #   beep = Sound.new("./assets/beep.wav")
+  #   # Do some stuff...
+  #   beep.unload
+  #
+  # @return [nil]
+  def unload
+    # mrb_Sound_unload
+    # src/mruby_integration/models/sound.cpp
+    nil
+  end
+
   # Returns the frame_count for the {Sonud} object.
+  #
+  # @example Basic usage
+  #   beep = Sound.new("./assets/beep.wav")
+  #   puts beep.frame_count #=> 96000
   #
   # @return [Integer]
   def frame_count